LuaStack is no longer automatically convertible to its topmost value
This was a neat idea at the time, but I think it adds a lot of mystery to
how LuaStack operates in practice. Furthermore, since the reference support
and LuaValue have both gotten more stable, there's not as much of a need
to keep LuaStack around as a faux variable.