Rewrite of LuaUserdata handling
The old implementation had a very clunky model for handling userdata. Each
userdata corresponded to a separate C++ object, that wrapped an underlying
C++ object. Both objects had to be wrapped in shared_ptr's to be safe. This
made using the userdata quite verbose (to the point where it was hard to
keep track of the nesting and indirection).

This model solves the problem by doing a few things:

1. It decouples userdata creation from metatable creation
2. Userdata objects are no longer necessary, as metatable creation can be
done using closures
3. All Lua objects are wrapped in a LuaUserdata object, that contains a
shared_ptr<void> to its data and a QString that describes its type.
4. Garbage collection on LuaUserdata is always done internally