LuaStack prefers >> and references over to()
This allows external clients to override >> and immediately be able to
use those types in LuaStack. This is the primary benefit.

Using references over pointers seems a bit more C++ish. iostream stuff
uses references, so I figure I would as well. In practice, the code is
easier to read.