Added LuaTableAccessible
This is a symbolic reference that will use another accessible to resolve
itself. This is very convenient for constructs like:

lua["Foo"]["Bar"] = myBarFunction;

While there are uses for hard references, it's very rare that we want to
refer to a global and then save it to some other value. Of course, it's
still very possible to do so.