LuaAccessibles can now be used from the stack
LuaValue is now a templated class that can store an accessible of
arbitrary type (as long as it looks like an LuaAccessible ;) ). This
means that global and table accesses don't need to be stack-created.

The reason why I didn't do this before is because references need a
reference-counting mechanism of some sort. At the time, I was compelled
to introduce shared pointers to enable this reference counting
mechanism. This polluted every creation of LuaValue for no good reason.

LuaReferenceAccessible is still reference-counted, but its shared_ptr is
internal to the class. This allows us to treat this accessible like any
other value, while still having the same behavior as before.

There's other changes here, among them:

* Moved several functions and method definitions into cpp files
* Merged LuaAccessible.hpp into LuaStack
* lua_State* is now accessible via a public method. This lets people
access this data when they need to. Of course, they do so at their own
risk.
* Removed some uses of 'using std::string' and such
* Added typedefs for LuaGlobal, LuaReference, and LuaTableValue