Allow overloads of as<Sink>
This is in preparation for allowing userdata to be passed by reference.
It's also for quality-of-life; keeping my Lua-specific overloads in a
named class allows me to have much better error messages. It also
simplifies the definition of user overloads, since they don't need to
worry about handling the LuaIndex.