Greatly improves how Lua-to-C++ method invocation works
The main purpose of this patch was to simplify how new userdata types
can be introduced. The old version required users to manually overload
operator>> for all types that they intended to handle. This typically
meant that many operators>> were either omitted or implemented poorly.

Instead, I'm using a much more sophisticated approach that provides equal
or better flexibility as the old one, with the massive benefit that
userdata types can be introduced with only two overloads:

template <>
UserdataName<YourType>
{
    constexpr static const char* value = "YourType";
};

As well as operator<<, which behaves as before (this requirement will be
eliminated shortly).

As a side-effect, this rewrite has also widened the number of accepted
types dramatically by allowing references to be passed in. It also allows
direct access to userdata objects via reference, which wasn't supported
previously.