Grabbing with a parent is now okay
The utility gained here is being able to have the parent-stack
protection while still being able to do fancy things. Normally, I
wouldn't like allowing this (hence the old implementation), but these
methods are so low-level and esoteric that there's not really a safe way
to implement them.

I also made unlock() not check to see if the stack was changed. We trust
LuaStack itself to clean up stack arguments, so if the stack does have
values left over from the child, they were meant for the parent.