Removed use of returning LuaStack& where unnecessary
I initially wanted to remove every use of this idiom, but there's a few
where it's appropriate to chain. Specifically global() and get() were
used in chains (e.g. get("foo").type()) and I think this is reasonable.
For the most part, though, it just dirties things up without providing a
benefit.