Catch a common bug earlier and more explicitly
When fiddling with how we convert C++ functions to Lua, we usually end
up trying to treat void(LuaStack&) as an arbitrary function, instead of
one that can handle invocation natively. This ends up failing at compile
time since there's no way to convert a Lua value to a LuaStack&, which
is good. The bad thing is that the error message usually wastes 5
minutes of my time every time I see it. So instead, this patch catches
that problem as directly as possible.