Minimized use of operator<<
This causes problems due to how template resolution is done. Since
operator<< was defined so early, it would only use those lua::push
definitions within LuaStack.hpp. As a result, anything defined outside
of there would be ignored and problems would result.

There's still the possibility of things like this causing issues. For
example, function invocation is done within LuaStack.hpp, so it's
possible that returning an unexpected value from there would break
things.

Perhaps the best way to fix this long-term is reorganize the
files such that all of the lua::push definitions can be done before any
of the lua::push-using code is included. This would work, but it would
be a major pain.