Added manual memory management override
Qt manages a widget's memory internally, which means we can't use shared
pointers lest we end up with a double delete. I don't want to disable
memory protection globally - since it's a very useful feature - and I
don't want to use light userdata hacks.

The solution is this: an override on a per-userdata basis. If a userdata
has this override, then it is allowed to pass through the memory
protection checks. This behavior is similar to a cast in C++: you should
never need it unless you're wiring ugly things together.