Support push for LuaValue
This was causing problems due to the way function overloading is handled
in C++. Class specialization seems to resolve later in the process,
which I take advantage of to ensure LuaValue always works properly.

If you're curious what the problem was (it's fairly hard to explain),
just run the test case without the include/* additions.