LuaUserdata throws when it is accessed with no data
This only applies in practice to stack-based userdata. When this
userdata falls out of scope, but Lua itself retained a reference to it,
it would remain as a zombie. It would suddenly have no properties and
would fail silently. This is quite confusing to deal with, so I've
decided to just throw as soon as access is made.

One problem is that Lua itself has no way of knowing whether a userdata
is valid or not; only the exception will serve as an indicator that the
userdata has perished. In practice, I don't think this will be too much
of a problem, as long as the error message is specific enough.