Use lua.h instead of lua.hpp
AFAIK, we need to use lua.h instead of lua.hpp because of how we
treat exceptions. Lua needs to be built as C++ in order to properly
throw exceptions, but this will mangle the names. lua.hpp externs the
internal defintions, which confuses the linker.

I don't believe that Fedora works in this case. I think Lua is built as
C and the stack is unwound erroneously, but I haven't yet confirmed this.