Significant reworking of how stack positions are handled
They are now LuaStack-local, rather than absolute. The conversion to
absolute positions is done internal to LuaStack. checkPos() was changed
to absPos() to reflect this change, and all positions are now const to
ensure a double conversion can't easily take place (since absPos() has
no way of knowing that a position is absolute)