Simplified how C++ functions are converted to lua::functions
This is more of a stylistic cleanup than a performance win, but I
removed the need for a separate wrapping object. Now only a templated
function is passed, along with the function userdata. However, since
this stuff is used rather often in Lua-cxx, it's nice to have it be
as flat as possible.