Simplified how const char* is converted
The old way would cause issues if the string was garbage collected
or moved in any way. This is the reason why Lua forces non-strings
to be converted to strings in-place on the stack. I've never liked
the way this is done, but the alternative of failing would be
inconsistent with Lua's approach, so I will just defer to them.