Removed link, index dependencies from stack.hpp
This forces a change in the Get, Store, Push, and Metatable structs, but
frees up those definitions to just use lua_State* and int. It also means
that lua::index and lua::link become just as optional as lua::reference
and lua::global. This simplifies the interdependencies within stack
dramatically.

Of course, it also breaks every binded type.