Luacxx no longer provides a default metatable for types
This was an extremely beneficial feature to have when Metatables were
difficult to create and uses of Luacxx were simple. However, the Qt binding
revealed that bindings are quite difficult to get right when default
Metatables are being created. The problems would manifest as some objects
that should be recognizable types showing up instead as anonymous userdata.
This is because a translation unit did not observe the correct
specialization.

Fixing these problems were quite difficult, as the missing include was
quite difficult to track down without the aid of the compiler. Now, missing
includes will cause a static_assert at the exact location of the improper
invocation. This far outweighs the inconsistency of default metatables.

Regardless, it is still possible to use default metatables by defining
LUACXX_AUTO_METATABLE where appropriate.